feat(money): dinheiro em Decimal (auditoria #4) — exatidão financeira
Substitui f64 por rust_decimal::Decimal em todo valor monetário do domínio
(preço, custo, total, subtotal, desconto, saldo, valores de caixa/carteira/
assinatura). Quantidades (peso/unidades) seguem f64 — não são dinheiro.

- core: novo módulo money (qty/round2/to_cents/from_db_f64); discount.rs e os
  cálculos de order/cash/wallet/coupon/subscription/finance/plan reescritos em
  Decimal. Testes de dinheiro (desconto, cupom, resumo de caixa) verdes (80).
- server (Postgres): colunas de dinheiro → NUMERIC(14,2) (sqlx Decimal nativo);
  DTOs do catálogo convertem Decimal→f64 na fronteira pública (web intocado, §11).
- desktop (SQLite, sem tipo decimal): cache em REAL, conversão na fronteira do
  repo (from_db_f64 c/ round2 na leitura, to_f64 no bind); cálculos exatos.
- serde de Decimal uniforme nos 3 crates → sync consistente. Bancos limpos
  (fresh start), sem clientes antigos a quebrar.

clippy --workspace limpo (0 warnings); cargo test -p letaf-core 80/0; app desktop
sobe sem panic com SQLite recriado.

+++ core/src/addon/model.rs
@@ -1,4 +1,5 @@
 use serde::{Deserialize, Serialize};
+use rust_decimal::Decimal;
 use uuid::Uuid;
 
 use crate::entity::BaseFields;
@@ -20,7 +21,7 @@ pub struct Addon {
     pub name: String,
     /// Acréscimo em R$ somado ao preço base do produto. `0.0` é válido
     /// (ex.: "Sem cebola" gratuito).
-    pub price: f64,
+    pub price: Decimal,
     /// Ordem dentro do grupo (asc).
     #[serde(default)]
     pub sort_order: i32,
@@ -35,7 +36,7 @@ fn default_true() -> bool {
 }
 
 impl Addon {
-    pub fn new(company_id: Uuid, group_id: Uuid, name: String, price: f64) -> Self {
+    pub fn new(company_id: Uuid, group_id: Uuid, name: String, price: Decimal) -> Self {
         Self {
             base: BaseFields::new(company_id),
             group_id,
